добавил пакеты для менторства
