+++ src/main/java/school/faang/user_service/dto/mentorship/MentorshipFilterDto.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.dto;
+package school.faang.user_service.dto.mentorship;
 
 import school.faang.user_service.entity.RequestStatus;
