+++ src/main/java/school/faang/user_service/dto/mentorship/MentorshipResponseDto.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.dto;
+package school.faang.user_service.dto.mentorship;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
