+++ src/main/java/school/faang/user_service/filter/mentorship/StatusFilter.java
@@ -1,11 +1,12 @@
-package school.faang.user_service.filter;
+package school.faang.user_service.filter.mentorship;
 
-import school.faang.user_service.dto.MentorshipFilterDto;
+import school.faang.user_service.dto.mentorship.MentorshipFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
+import school.faang.user_service.filter.mentorship.MentorshipFilter;
 
 import java.util.stream.Stream;
 
-public class StatusFilter implements MentorshipFilter{
+public class StatusFilter implements MentorshipFilter {
     @Override
     public boolean isApplicable(MentorshipFilterDto filterDto) {
         return filterDto.status() != null;
