+++ src/main/java/school/faang/user_service/rest/ExceptionHandlerControllerAdvice.java
@@ -7,10 +7,8 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.server.ResponseStatusException;
-import school.faang.user_service.dto.ErrorResponseDto;
+import school.faang.user_service.dto.mentorship.ErrorResponseDto;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 @RestControllerAdvice
