+++ src/main/java/school/faang/user_service/service/MentorshipRequestService.java
@@ -3,12 +3,12 @@ package school.faang.user_service.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import school.faang.user_service.dto.MentorshipFilterDto;
-import school.faang.user_service.dto.MentorshipRequestDto;
-import school.faang.user_service.dto.MentorshipResponseDto;
+import school.faang.user_service.dto.mentorship.MentorshipFilterDto;
+import school.faang.user_service.dto.mentorship.MentorshipRequestDto;
+import school.faang.user_service.dto.mentorship.MentorshipResponseDto;
 import school.faang.user_service.entity.MentorshipRequest;
-import school.faang.user_service.filter.MentorshipFilter;
-import school.faang.user_service.mapper.MentorshipResponseMapper;
+import school.faang.user_service.filter.mentorship.MentorshipFilter;
+import school.faang.user_service.mapper.mentorship.MentorshipResponseMapper;
 import school.faang.user_service.repository.mentorship.MentorshipRepository;
 import school.faang.user_service.repository.mentorship.MentorshipRequestRepository;
 import school.faang.user_service.validation.MentorshipValidator;
